+++ packages/language-ruby/lib/main.js
@@ -7,17 +7,50 @@ exports.activate = function() {
       return node.lastChild.text;
     },
     content(node) {
-      return node;
-    }
+      return node.descendantsOfType('heredoc_content')
+    },
+    // coverShallowerScopes: true
   });
 
   atom.grammars.addInjectionPoint('source.ruby', {
     type: 'regex',
     language() {
-      return 'regex';
+      return 'rb-regex';
     },
     content(node) {
       return node;
-    }
+    },
+    includeChildren: true,
+    // coverShallowerScopes: false
+  });
+
+  const TODO_PATTERN = /\b(TODO|FIXME|CHANGED|XXX|IDEA|HACK|NOTE|REVIEW|NB|BUG|QUESTION|COMBAK|TEMP|DEBUG|OPTIMIZE|WARNING)\b/;
+  const HYPERLINK_PATTERN = /\bhttps?:/
+
+  atom.grammars.addInjectionPoint('source.ruby', {
+    type: 'comment',
+    language: (node) => {
+      return TODO_PATTERN.test(node.text) ? 'todo' : null;
+    },
+    content: (node) => node,
+    languageScope: null
+  });
+
+  atom.grammars.addInjectionPoint('source.ruby', {
+    type: 'comment',
+    language: (node) => {
+      return HYPERLINK_PATTERN.test(node.text) ? 'hyperlink' : null;
+    },
+    content: (node) => node,
+    languageScope: null
+  });
+
+  atom.grammars.addInjectionPoint('source.ruby', {
+    type: 'string_content',
+    language: (node) => {
+      return HYPERLINK_PATTERN.test(node.text) ? 'hyperlink' : null;
+    },
+    content: (node) => node,
+    languageScope: null
   });
 };