+++ spec/grammar-registry-spec.js
@@ -8,13 +8,21 @@ const TreeSitterGrammar = require('../src/tree-sitter-grammar');
 const SecondMate = require('second-mate');
 const { OnigScanner } = SecondMate;
 
+// Expects one of `textmate`, `node-tree-sitter`, or `wasm-tree-sitter`.
+function setConfigForLanguageMode(mode, options = {}) {
+  let useTreeSitterParsers = mode !== 'textmate';
+  let useExperimentalModernTreeSitter = mode === 'wasm-tree-sitter';
+  atom.config.set('core.useTreeSitterParsers', useTreeSitterParsers, options);
+  atom.config.set('core.useExperimentalModernTreeSitter', useExperimentalModernTreeSitter, options);
+}
+
 describe('GrammarRegistry', () => {
   let grammarRegistry;
 
   beforeEach(async () => {
     await SecondMate.ready
     grammarRegistry = new GrammarRegistry({ config: atom.config });
-    expect(subscriptionCount(grammarRegistry)).toBe(1);
+    expect(subscriptionCount(grammarRegistry)).toBe(2);
   });
 
   describe('.assignLanguageMode(buffer, languageId)', () => {
@@ -88,10 +96,8 @@ describe('GrammarRegistry', () => {
   });
 
   describe('.grammarForId(languageId)', () => {
-    it('returns a text-mate grammar when `core.useTreeSitterParsers` is false', () => {
-      atom.config.set('core.useTreeSitterParsers', false, {
-        scopeSelector: '.source.js'
-      });
+    it('returns a text-mate grammar when config is set to `textmate`', () => {
+      setConfigForLanguageMode('textmate', { scopeSelector: '.source.js' });
 
       grammarRegistry.loadGrammarSync(
         require.resolve('language-javascript/grammars/javascript.cson')
@@ -110,8 +116,8 @@ describe('GrammarRegistry', () => {
       expect(grammarRegistry.grammarForId('javascript')).toBe(undefined);
     });
 
-    it('returns a tree-sitter grammar when `core.useTreeSitterParsers` is true', () => {
-      atom.config.set('core.useTreeSitterParsers', true);
+    it('returns a tree-sitter grammar when language is set to node-tree-sitter', () => {
+      setConfigForLanguageMode('node-tree-sitter');
 
       grammarRegistry.loadGrammarSync(
         require.resolve('language-javascript/grammars/javascript.cson')
@@ -174,7 +180,7 @@ describe('GrammarRegistry', () => {
     });
 
     it("updates the buffer's grammar when a more appropriate text-mate grammar is added for its path", async () => {
-      atom.config.set('core.useTreeSitterParsers', false);
+      setConfigForLanguageMode('textmate');
 
       const buffer = new TextBuffer();
       expect(buffer.getLanguageMode().getLanguageId()).toBe(null);
@@ -196,7 +202,7 @@ describe('GrammarRegistry', () => {
     });
 
     it("updates the buffer's grammar when a more appropriate tree-sitter grammar is added for its path", async () => {
-      atom.config.set('core.useTreeSitterParsers', true);
+      setConfigForLanguageMode('node-tree-sitter');
 
       const buffer = new TextBuffer();
       expect(buffer.getLanguageMode().getLanguageId()).toBe(null);
@@ -216,6 +222,68 @@ describe('GrammarRegistry', () => {
       );
       expect(buffer.getLanguageMode().grammar).toBe(treeSitterGrammar);
     });
+
+    it("updates the buffer's grammar when a more appropriate new-tree-sitter grammar is added for its path and the user has opted into new-tree-sitter", async () => {
+      setConfigForLanguageMode('wasm-tree-sitter');
+
+      const buffer = new TextBuffer();
+      expect(buffer.getLanguageMode().getLanguageId()).toBe(null);
+
+      buffer.setPath('test.js');
+      grammarRegistry.maintainLanguageMode(buffer);
+
+      const treeSitterGrammar = grammarRegistry.loadGrammarSync(
+        require.resolve(
+          'language-javascript/grammars/tree-sitter-javascript.cson'
+        )
+      );
+      expect(buffer.getLanguageMode().grammar).toBe(treeSitterGrammar);
+
+      // TODO: Why doesn't this path resolution work like the one above?
+      const modernTreeSitterGrammar = grammarRegistry.loadGrammarSync(
+        require.resolve(
+          '../packages/language-javascript/grammars/tree-sitter-2-javascript.cson'
+        )
+      );
+      expect(buffer.getLanguageMode().grammar).toBe(modernTreeSitterGrammar);
+      grammarRegistry.loadGrammarSync(
+        require.resolve('language-javascript/grammars/javascript.cson')
+      );
+      expect(buffer.getLanguageMode().grammar).toBe(modernTreeSitterGrammar);
+    });
+
+    it("updates the buffer's grammar by ignoring a new-tree-sitter grammar if the user has NOT opted into new-tree-sitter", async () => {
+      setConfigForLanguageMode('node-tree-sitter');
+
+      const buffer = new TextBuffer();
+      expect(buffer.getLanguageMode().getLanguageId()).toBe(null);
+
+      buffer.setPath('test.js');
+      grammarRegistry.maintainLanguageMode(buffer);
+
+      const textmateGrammar = grammarRegistry.loadGrammarSync(
+        require.resolve('language-javascript/grammars/javascript.cson')
+      );
+      expect(buffer.getLanguageMode().grammar).toBe(textmateGrammar);
+
+      // TODO: Why doesn't this path resolution work like the one above?
+      const modernTreeSitterGrammar = grammarRegistry.loadGrammarSync(
+        require.resolve(
+          '../packages/language-javascript/grammars/tree-sitter-2-javascript.cson'
+        )
+      );
+
+      expect(buffer.getLanguageMode().grammar).toBe(textmateGrammar);
+
+      const treeSitterGrammar = grammarRegistry.loadGrammarSync(
+        require.resolve(
+          'language-javascript/grammars/tree-sitter-javascript.cson'
+        )
+      );
+
+      expect(buffer.getLanguageMode().grammar).toBe(treeSitterGrammar);
+    });
+
 
     it('can be overridden by calling .assignLanguageMode', () => {
       const buffer = new TextBuffer();
@@ -303,16 +371,16 @@ describe('GrammarRegistry', () => {
 
       const disposable = grammarRegistry.maintainLanguageMode(buffer);
       expect(retainedBufferCount(grammarRegistry)).toBe(1);
-      expect(subscriptionCount(grammarRegistry)).toBe(3);
+      expect(subscriptionCount(grammarRegistry)).toBe(4);
 
       buffer.destroy();
       expect(retainedBufferCount(grammarRegistry)).toBe(0);
-      expect(subscriptionCount(grammarRegistry)).toBe(1);
+      expect(subscriptionCount(grammarRegistry)).toBe(2);
       expect(buffer.emitter.getTotalListenerCount()).toBe(0);
 
       disposable.dispose();
       expect(retainedBufferCount(grammarRegistry)).toBe(0);
-      expect(subscriptionCount(grammarRegistry)).toBe(1);
+      expect(subscriptionCount(grammarRegistry)).toBe(2);
     });
 
     it('does not retain the buffer when the grammar registry is destroyed', () => {
@@ -323,12 +391,12 @@ describe('GrammarRegistry', () => {
 
       grammarRegistry.maintainLanguageMode(buffer);
       expect(retainedBufferCount(grammarRegistry)).toBe(1);
-      expect(subscriptionCount(grammarRegistry)).toBe(3);
+      expect(subscriptionCount(grammarRegistry)).toBe(4);
 
       grammarRegistry.clear();
 
       expect(retainedBufferCount(grammarRegistry)).toBe(0);
-      expect(subscriptionCount(grammarRegistry)).toBe(1);
+      expect(subscriptionCount(grammarRegistry)).toBe(2);
       expect(buffer.emitter.getTotalListenerCount()).toBe(0);
     });
   });
@@ -545,10 +613,8 @@ describe('GrammarRegistry', () => {
     });
 
     describe('tree-sitter vs text-mate', () => {
-      it('favors a text-mate grammar over a tree-sitter grammar when `core.useTreeSitterParsers` is false', () => {
-        atom.config.set('core.useTreeSitterParsers', false, {
-          scopeSelector: '.source.js'
-        });
+      it('favors a text-mate grammar over a tree-sitter grammar when config is set to textmate', () => {
+        setConfigForLanguageMode('textmate', { scopeSelector: '.source.js' });
 
         grammarRegistry.loadGrammarSync(
           require.resolve('language-javascript/grammars/javascript.cson')
@@ -564,8 +630,9 @@ describe('GrammarRegistry', () => {
         expect(grammar instanceof SecondMate.Grammar).toBe(true);
       });
 
-      it('favors a tree-sitter grammar over a text-mate grammar when `core.useTreeSitterParsers` is true', () => {
-        atom.config.set('core.useTreeSitterParsers', true);
+      it('favors a tree-sitter grammar over a text-mate grammar when config is set', () => {
+        setConfigForLanguageMode('node-tree-sitter');
+
 
         grammarRegistry.loadGrammarSync(
           require.resolve('language-javascript/grammars/javascript.cson')
@@ -581,7 +648,8 @@ describe('GrammarRegistry', () => {
       });
 
       it('only favors a tree-sitter grammar if it actually matches in some way (regression)', () => {
-        atom.config.set('core.useTreeSitterParsers', true);
+        setConfigForLanguageMode('node-tree-sitter');
+
         grammarRegistry.loadGrammarSync(
           require.resolve(
             'language-javascript/grammars/tree-sitter-javascript.cson'
@@ -595,7 +663,8 @@ describe('GrammarRegistry', () => {
 
     describe('tree-sitter grammars with content regexes', () => {
       it('recognizes C++ header files', () => {
-        atom.config.set('core.useTreeSitterParsers', true);
+        setConfigForLanguageMode('node-tree-sitter');
+
         grammarRegistry.loadGrammarSync(
           require.resolve('language-c/grammars/tree-sitter-c.cson')
         );
@@ -644,7 +713,8 @@ describe('GrammarRegistry', () => {
       });
 
       it('recognizes C++ files that do not match the content regex (regression)', () => {
-        atom.config.set('core.useTreeSitterParsers', true);
+        setConfigForLanguageMode('node-tree-sitter');
+
         grammarRegistry.loadGrammarSync(
           require.resolve('language-c/grammars/tree-sitter-c.cson')
         );
@@ -665,7 +735,7 @@ describe('GrammarRegistry', () => {
       });
 
       it('does not apply content regexes from grammars without filetype or first line matches', () => {
-        atom.config.set('core.useTreeSitterParsers', true);
+        setConfigForLanguageMode('node-tree-sitter');
         grammarRegistry.loadGrammarSync(
           require.resolve('language-c/grammars/tree-sitter-cpp.cson')
         );
@@ -683,7 +753,7 @@ describe('GrammarRegistry', () => {
       });
 
       it('recognizes shell scripts with shebang lines', () => {
-        atom.config.set('core.useTreeSitterParsers', true);
+        setConfigForLanguageMode('node-tree-sitter');
         grammarRegistry.loadGrammarSync(
           require.resolve('language-shellscript/grammars/shell-unix-bash.cson')
         );
@@ -713,7 +783,7 @@ describe('GrammarRegistry', () => {
         expect(grammar.name).toBe('Shell Script');
         expect(grammar instanceof TreeSitterGrammar).toBeTruthy();
 
-        atom.config.set('core.useTreeSitterParsers', false);
+        setConfigForLanguageMode('textmate');
         grammar = grammarRegistry.selectGrammar(
           'test.h',
           dedent`
@@ -727,7 +797,7 @@ describe('GrammarRegistry', () => {
       });
 
       it('recognizes JavaScript files that use Flow', () => {
-        atom.config.set('core.useTreeSitterParsers', true);
+        setConfigForLanguageMode('node-tree-sitter');
         grammarRegistry.loadGrammarSync(
           require.resolve(
             'language-javascript/grammars/tree-sitter-javascript.cson'
@@ -789,9 +859,11 @@ describe('GrammarRegistry', () => {
       await atom.packages.activatePackage('language-css');
       const grammar = atom.grammars.selectGrammar('foo.css');
       atom.grammars.removeGrammar(grammar);
-      expect(atom.grammars.selectGrammar('foo.css').name).not.toBe(
-        grammar.name
-      );
+      let newGrammar = atom.grammars.selectGrammar('foo.css');
+      expect(
+        grammar.name === newGrammar.name &&
+          grammar.constructor.name === newGrammar.constructor.name
+      ).toBe(false);
     });
   });
 
@@ -807,7 +879,7 @@ describe('GrammarRegistry', () => {
     };
 
     beforeEach(() => {
-      atom.config.set('core.useTreeSitterParsers', true);
+      setConfigForLanguageMode('node-tree-sitter');
     });
 
     it('adds an injection point to the grammar with the given id', async () => {